+++ intelmq/tests/bots/experts/cymru_whois/test_expert.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import json
 import unittest
 
 import intelmq.lib.test as test
@@ -40,20 +41,6 @@ EXAMPLE_6TO4_INPUT = {"__type": "Event",
                  "source.ip": "2002:3ee0:3972:0001::1",
                  "time.observation": "2015-01-01T00:00:00+00:00",
                  }
-EXAMPLE_6TO4_OUTPUT = {"__type": "Event",
-                  "source.ip": "2002:3ee0:3972:0001::1",
-                  "source.network": "2002::/16",
-                  "source.asn": 1103,
-                  "source.as_name": "SURFNET-NL SURFnet, The Netherlands, NL",
-                  "time.observation": "2015-01-01T00:00:00+00:00",
-                  }
-EXAMPLE_6TO4_OUTPUT_1 = {"__type": "Event",
-                  "source.ip": "2002:3ee0:3972:0001::1",
-                  "source.network": "2002::/16",
-                  "source.asn": 6939,
-                  "source.as_name": "HURRICANE, US",
-                  "time.observation": "2015-01-01T00:00:00+00:00",
-                  }
 OVERWRITE_OUT = {"__type": "Event",
                   "source.ip": "78.104.144.2",
                   "source.geolocation.cc": "AA",
@@ -98,20 +85,21 @@ class TestCymruExpertBot(test.BotTestCase, unittest.TestCase):
     def test_6to4_result(self):
         """
         Test the whois for an IPv6 to IPv4 network range.
-        The result can vary, so we test for two possible expected results.
+        The result can vary, so we only tests if values exist.
         """
         self.input_message = EXAMPLE_6TO4_INPUT
         self.run_bot()
-        try:
-            self.assertMessageEqual(0, EXAMPLE_6TO4_OUTPUT)
-        except AssertionError:
-            self.assertMessageEqual(0, EXAMPLE_6TO4_OUTPUT_1)
+        actual = json.loads(self.get_output_queue()[0])
+        self.assertDictContainsSubset(EXAMPLE_6TO4_INPUT, actual)
+        self.assertIn("source.asn", actual)
+        self.assertIn("source.as_name", actual)
+        self.assertIn("source.network", actual)
 
     def test_overwrite(self):
         self.input_message = EXAMPLE_INPUT.copy()
         self.input_message["source.geolocation.cc"] = "AA"
         self.input_message["source.registry"] = "LACNIC"
-        self.run_bot(parameters={'overwrite' : False})
+        self.run_bot(parameters={'overwrite': False})
         self.assertMessageEqual(0, OVERWRITE_OUT)